requirements.txt
fritzconnection
python-dotenv
pyyaml
